Move some global func to LedgerState, fix build
