+++ src/lib.rs
@@ -40,7 +40,6 @@ pub mod testing;
 pub mod transactions;
 pub mod txn_builder;
 
-use crate::sparse_merkle_tree::SparseMerkleTree;
 pub use crate::{
     assets::{Asset, AssetEditor, Assets, Icon, MintInfo},
     txn_builder::RecordAmount,
@@ -54,7 +53,7 @@ use crate::{
     encryption::Cipher,
     events::{EventIndex, EventSource, LedgerEvent},
     hd::KeyTree,
-    key_scan::{receive_history_entry, BackgroundKeyScan, ScanOutputs},
+    key_scan::{receive_history_entry, ScanOutputs},
     ledger_state::{LedgerState, LedgerStateStore},
     loader::KeystoreLoader,
     persistence::AtomicKeystoreStorage,
@@ -84,28 +83,20 @@ use jf_cap::{
     },
     mint::MintNote,
     structs::{
-        AssetCode, AssetCodeSeed, AssetDefinition, AssetPolicy, FreezeFlag, Nullifier,
-        ReceiverMemo, RecordCommitment, RecordOpening,
+        AssetCode, AssetDefinition, AssetPolicy, FreezeFlag, Nullifier, ReceiverMemo,
+        RecordCommitment, RecordOpening,
     },
     transfer::TransferNote,
-    MerkleLeafProof, MerklePath, MerkleTree, TransactionNote, VerKey,
+    MerklePath, MerkleTree, TransactionNote, VerKey,
 };
 use jf_primitives::aead;
-use key_set::ProverKeySet;
 use primitive_types::U256;
 use rand_chacha::{rand_core::SeedableRng, ChaChaRng};
-use reef::{
-    traits::{
-        Block as _, NullifierSet as _, Transaction as _, TransactionKind as _,
-        ValidationError as _, Validator as _,
-    },
-    TransactionKind, *,
-};
+use reef::{traits::Transaction as _, TransactionKind, *};
 use serde::{de::DeserializeOwned, Serialize};
-use snafu::{ResultExt, Snafu};
+use snafu::Snafu;
 use std::collections::HashMap;
 use std::convert::TryFrom;
-use std::iter::repeat;
 use std::sync::Arc;
 use std::time::Duration;
 
@@ -377,44 +368,6 @@ impl<
     }
 }
 
-pub fn balance<
-    'a,
-    L: Ledger + DeserializeOwned + Serialize,
-    Meta: Serialize + DeserializeOwned + Send,
->(
-    state: &mut LedgerState<'a, L>,
-    model: &mut KeystoreModel<'a, L, impl KeystoreBackend<'a, L>, Meta>,
-    asset: &AssetCode,
-    frozen: FreezeFlag,
-) -> U256 {
-    let mut balance = U256::zero();
-    for address in model.sending_accounts.iter_pub_keys() {
-        balance += balance_breakdown(&mut state, &mut model, &address, asset, frozen);
-    }
-    balance
-}
-
-pub fn balance_breakdown<
-    'a,
-    L: Ledger + DeserializeOwned + Serialize,
-    Meta: Serialize + DeserializeOwned + Send,
->(
-    state: &mut LedgerState<'a, L>,
-    model: &mut KeystoreModel<'a, L, impl KeystoreBackend<'a, L>, Meta>,
-    address: &UserAddress,
-    asset: &AssetCode,
-    frozen: FreezeFlag,
-) -> U256 {
-    let spendable = model.records.get_spendable::<L>(asset, address, frozen);
-    if let Some(records) = spendable {
-        records
-            .filter(move |record| !record.on_hold(state.block_height()))
-            .fold(U256::zero(), |sum, record| sum + record.amount())
-    } else {
-        U256::zero()
-    }
-}
-
 // Inform the Transactions database that we have received memos for the given record UIDs. Return a list of
 // the transactions that are completed as a result.
 pub fn received_memos<L: Ledger>(
@@ -434,544 +387,42 @@ pub fn received_memos<L: Ledger>(
     completed
 }
 
-pub async fn handle_event<
-    'a,
-    L: Ledger + DeserializeOwned + Serialize,
-    Meta: Serialize + DeserializeOwned + Send,
->(
-    state: &mut LedgerState<'a, L>,
-    model: &mut KeystoreModel<'a, L, impl KeystoreBackend<'a, L>, Meta>,
-    event: LedgerEvent<L>,
-    source: EventSource,
-) -> Result<EventSummary<L>, KeystoreError<L>> {
-    state.increment_now();
-    let mut summary = EventSummary::default();
-    match event {
-        LedgerEvent::Commit {
-            block,
-            block_id,
-            state_comm,
-        } => {
-            // Don't trust the network connection that provided us this event; validate it
-            // against our local mirror of the ledger and bail out if it is invalid.
-            let mut uids = match state.validator().validate_and_apply(block.clone()) {
-                Ok(uids) => {
-                    if state_comm != state.validator().commit() {
-                        // Received a block which validates, but our state commitment does not
-                        // match that of the event source. Since the block validates, we will
-                        // accept it, but this must indicate that the event source is lying or
-                        // mistaken about the state commitment. This would be a good time to
-                        // switch to a different query server or something, but for now we'll
-                        // just log the problem.
-                        tracing::error!("received valid block with invalid state commitment");
-                    }
-
-                    // Get a list of new uids and whether we want to remember them in our record
-                    // Merkle tree. Initially, set `remember` to false for all uids, to maximize
-                    // sparseness. If any of the consumers of this block (for example, the
-                    // viewer component, or the owner of this keystore) care about a uid, they
-                    // will set its `remember` flag to true.
-                    uids.into_iter().map(|uid| (uid, false)).collect::<Vec<_>>()
-                }
-                Err(val_err) => {
-                    //todo !jeb.bearer handle this case more robustly. If we get here, it means
-                    // the event stream has lied to us, so recovery is quite tricky and may
-                    // require us to fail over to a different query service.
-                    panic!("received invalid block: {:?}, {:?}", block, val_err);
-                }
-            };
-
-            // Update our full copies of sparse validator data structures to be consistent with
-            // the validator state.
-            for txn in block.txns() {
-                let nullifiers = txn.input_nullifiers();
-                // Remove spent records.
-                for n in &nullifiers {
-                    if let Ok(record) = model.records.delete_by_nullifier::<L>(n) {
-                        state.forget_merkle_leaf(record.uid());
-                    }
-                }
-            }
-            // Insert new records.
-            state.append_merkle_leaves(
-                block
-                    .txns()
-                    .into_iter()
-                    .flat_map(|txn| txn.output_commitments()),
-            );
-            // Update nullifier set
-            let nullifier_proofs = block
-                .txns()
-                .into_iter()
-                .flat_map(|txn| txn.proven_nullifiers())
-                .collect::<Vec<_>>();
-            if state.nullifiers().multi_insert(&nullifier_proofs).is_err() {
-                //todo !jeb.bearer handle this case more robustly. If we get here, it means the
-                // event stream has lied to us, so recovery is quite tricky and may require us
-                // to fail over to a different query service.
-                panic!("received block with invalid nullifier proof");
-            }
-
-            for (txn_id, txn) in block.txns().into_iter().enumerate() {
-                // Split the uids corresponding to this transaction off the front of `uids`.
-                let mut this_txn_uids = uids;
-                uids = this_txn_uids.split_off(txn.output_len());
-                assert_eq!(this_txn_uids.len(), txn.output_len());
-
-                // If this transaction contains record openings for all of its outputs,
-                // consider it retired immediately, do not wait for memos.
-                let retired = txn.output_openings().is_some();
-
-                // Add the spent nullifiers to the summary. Map each nullifier to one of the
-                // output UIDs of the same transaction, so that we can tell when the memos
-                // arrive for the transaction which spent this nullifier (completing the
-                // transaction's life cycle) by looking at the UIDs attached to the memos.
-                // TODO !keyao Stop identifying transactions by input nullifier and instead use hashes.
-                // (https://github.com/SpectrumXYZ/cape/issues/275.)
-                if !txn.input_nullifiers().is_empty() {
-                    summary.spent_nullifiers.extend(
-                        txn.input_nullifiers()
-                            .into_iter()
-                            .zip(repeat(this_txn_uids[0].0)),
-                    );
-                    if retired {
-                        summary.retired_nullifiers.push(txn.input_nullifiers()[0]);
-                    }
-                }
-
-                // Different concerns within the keystore consume transactions in different ways.
-                // Now we give each concern a chance to consume this transaction, performing any
-                // processing they need to do and possibly setting the `remember` flag for
-                // output records they care about.
-                //
-                // This is a transaction we submitted and have been
-                // awaiting confirmation.
-                let mut self_published = false;
-                if let Some(pending) = clear_pending_transaction(
-                    state,
-                    model,
-                    &txn,
-                    Some((block_id, txn_id as u64, &mut this_txn_uids)),
-                )
-                .await
-                {
-                    let status = if retired {
-                        TransactionStatus::Retired
-                    } else {
-                        TransactionStatus::AwaitingMemos
-                    };
-                    summary.updated_txns.push((pending.uid().clone(), status));
-                    model
-                        .transactions
-                        .get_mut(pending.uid())
-                        .unwrap()
-                        .add_pending_uids(
-                            &this_txn_uids
-                                .iter()
-                                .zip(pending.memos().unwrap().memos.iter())
-                                .filter_map(|((uid, _), memo)| memo.as_ref().map(|_| *uid))
-                                .into_iter()
-                                .collect::<Vec<u64>>(),
-                        )
-                        .set_status(status)
-                        .save()
-                        .unwrap();
-                    model
-                        .backend
-                        .finalize(pending, Some((block_id, txn_id as u64)))
-                        .await;
-                    self_published = true;
-                }
-
-                // This is someone else's transaction but we can view it.
-                view_transaction(model, &txn, &mut this_txn_uids).await?;
-
-                // If this transaction has record openings attached, check if they are for us
-                // and add them immediately, without waiting for memos.
-                if let Err(err) = receive_attached_records(
-                    model,
-                    &txn,
-                    &mut this_txn_uids,
-                    !self_published,
-                    // Only add to history if we didn't send this same transaction
-                )
-                .await
-                {
-                    tracing::error!(
-                        "Error saving records attached to transaction {}:{}: {}",
-                        block_id,
-                        txn_id,
-                        err
-                    );
-                }
-
-                // Prune the record Merkle tree of records we don't care about.
-                for (uid, remember) in this_txn_uids {
-                    if !remember {
-                        state.forget_merkle_leaf(uid);
-                    }
-                }
-            }
-
-            // Some transactions may have expired when we stepped the validator state. Remove
-            // them from our pending transaction data structures.
-            //
-            // This maintains the invariant that everything in `pending_transactions` must
-            // correspond to an on-hold record, because everything which corresponds to a record
-            // whose hold just expired will be removed from the set now.
-            match model.transactions.remove_expired(state.block_height()) {
-                Ok(txns) => {
-                    for txn in txns {
-                        summary
-                            .updated_txns
-                            .push((txn.uid().clone(), TransactionStatus::Rejected));
-                        model.backend.finalize(txn, None).await;
-                    }
-                }
-                Err(err) => {
-                    println!(
-                        "Error removing expired transaction from storage.  Block: {},  Error: {}",
-                        block_id, err
-                    );
-                }
-            }
-        }
-        LedgerEvent::Memos {
-            outputs,
-            transaction,
-        } => {
-            let completed =
-                received_memos(outputs.iter().map(|info| info.2), &mut model.transactions);
-            let self_published = !completed.is_empty();
-            summary.updated_txns.extend(
-                completed
-                    .into_iter()
-                    .map(|txn_uid| (txn_uid, TransactionStatus::Retired))
-                    .collect::<Vec<_>>(),
-            );
-
-            summary
-                .received_memos
-                .extend(outputs.iter().map(|(memo, _, uid, _)| (memo.clone(), *uid)));
-            for key in &model
-                .sending_accounts
-                .iter_keys()
-                .collect::<Vec<UserKeyPair>>()
-            {
-                let records =
-                    try_open_memos(model, key, &outputs, transaction.clone(), !self_published)
-                        .await?;
-                if let Err(err) = add_records(state, model, key, records).await {
-                    tracing::error!("error saving received records: {}", err);
-                }
-            }
-        }
-        LedgerEvent::Reject { block, error } => {
-            for mut txn in block.txns() {
-                summary
-                    .rejected_nullifiers
-                    .append(&mut txn.input_nullifiers());
-                if let Some(pending) = clear_pending_transaction(state, model, &txn, None).await {
-                    // Try to resubmit if the error is recoverable.
-                    let uid = pending.uid();
-                    if error.is_bad_nullifier_proof() {
-                        if update_nullifier_proofs(state, model, &mut txn)
-                            .await
-                            .is_ok()
-                            && submit_elaborated_transaction(state, model, txn, None)
-                                .await
-                                .await
-                                .is_ok()
-                        {
-                            // The transaction has been successfully resubmitted. It is still in
-                            // the same state (pending) so we don't need to add it to
-                            // `updated_txns`.
-                        } else {
-                            // If we failed to resubmit, then the rejection is final.
-                            summary
-                                .updated_txns
-                                .push((uid.clone(), TransactionStatus::Rejected));
-                            model.backend.finalize(pending, None).await;
-                        }
-                    } else {
-                        summary
-                            .updated_txns
-                            .push((uid.clone(), TransactionStatus::Rejected));
-                        model.backend.finalize(pending, None).await;
-                    }
-                }
-            }
-        }
-    };
-
-    model.ledger_state_store.update_dynamic(state).await?;
-    Ok(summary)
-}
-
-async fn try_open_memos<
+/// Import an unverified asset.
+///
+/// Note that this function cannot be used to import verified assets. Verified assets can only be
+/// imported using [verify_assets], conditional on a signature check.
+pub fn import_asset<
     'a,
     L: Ledger + DeserializeOwned + Serialize,
     Meta: Serialize + DeserializeOwned + Send,
 >(
     model: &mut KeystoreModel<'a, L, impl KeystoreBackend<'a, L>, Meta>,
-    key_pair: &UserKeyPair,
-    memos: &[(ReceiverMemo, RecordCommitment, u64, MerklePath)],
-    transaction: Option<(u64, u64, TransactionHash<L>, TransactionKind<L>)>,
-    add_to_history: bool,
-) -> Result<Vec<(RecordOpening, u64, MerklePath)>, KeystoreError<L>> {
-    let mut records = Vec::new();
-    for (memo, comm, uid, proof) in memos {
-        if let Ok(record_opening) = memo.decrypt(key_pair, comm, &[]) {
-            if !record_opening.is_dummy() {
-                // If this record is for us (i.e. its corresponding memo decrypts under
-                // our key) and not a dummy, then add it to our owned records.
-                records.push((record_opening, *uid, proof.clone()));
-            }
-        }
-    }
-
-    if add_to_history && !records.is_empty() {
-        if let Some((_block_id, _txn_id, hash, kind)) = transaction {
-            add_receive_history(
-                model,
-                kind,
-                hash,
-                &records
-                    .iter()
-                    .map(|(ro, _, _)| ro.clone())
-                    .collect::<Vec<_>>(),
-            )
-            .await?;
-        }
-    }
-
-    Ok(records)
-}
-
-async fn receive_attached_records<
-    'a,
-    L: Ledger + DeserializeOwned + Serialize,
-    Meta: Serialize + DeserializeOwned + Send,
->(
-    model: &mut KeystoreModel<'a, L, impl KeystoreBackend<'a, L>, Meta>,
-    txn: &reef::Transaction<L>,
-    uids: &mut [(u64, bool)],
-    add_to_history: bool,
+    asset: Asset,
 ) -> Result<(), KeystoreError<L>> {
-    let records = txn.output_openings().into_iter().flatten().zip(uids);
-    let mut my_records = vec![];
-    for (ro, (uid, remember)) in records {
-        if let Ok(account_editor) = model.sending_accounts.get_mut(&ro.pub_key.address()) {
-            // If this record is for us, add it to the keystore and include it in the
-            // list of received records for created a received transaction history
-            // entry.
-            *remember = true;
-            // Add the asset type if it is not already in the asset library.
-            model.assets.create(ro.asset_def.clone(), None)?;
-            // Mark the account receiving the records used.
-            let account = account_editor.set_used().save()?;
-            // Add the record.
-            model.records.create::<L>(
-                *uid,
-                ro.clone(),
-                account.key().nullify(
-                    ro.asset_def.policy_ref().freezer_pub_key(),
-                    *uid,
-                    &RecordCommitment::from(&ro),
-                ),
-            )?;
-            my_records.push(ro);
-        } else if let Ok(account_editor) = model
-            .freezing_accounts
-            .get_mut(ro.asset_def.policy_ref().freezer_pub_key())
-        {
-            // If this record is not for us, but we can freeze it, then this
-            // becomes like an view. Add the record to our collection of freezable
-            // records, but do not include it in the history entry.
-            *remember = true;
-            // Add the asset type if it is not already in the asset library.
-            model.assets.create(ro.asset_def.clone(), None)?;
-            // Mark the freezing account which is tracking the record used.
-            let account = account_editor.set_used().save()?;
-            // Add the record.
-            model.records.create::<L>(
-                *uid,
-                ro.clone(),
-                account
-                    .key()
-                    .nullify(&ro.pub_key.address(), *uid, &RecordCommitment::from(&ro)),
-            )?;
-        }
-    }
-
-    if add_to_history && !my_records.is_empty() {
-        add_receive_history(model, txn.kind(), txn.hash().clone(), &my_records).await?;
-    }
-
+    assert!(!asset.verified());
+    model
+        .assets
+        .create_internal(
+            asset.definition().clone(),
+            asset.mint_info(),
+            asset.verified(),
+        )?
+        .update_internal(asset)?
+        .save()?;
     Ok(())
 }
 
-async fn add_receive_history<
+/// Load a verified asset library with its trusted signer.
+pub fn verify_assets<
     'a,
     L: Ledger + DeserializeOwned + Serialize,
     Meta: Serialize + DeserializeOwned + Send,
 >(
     model: &mut KeystoreModel<'a, L, impl KeystoreBackend<'a, L>, Meta>,
-    kind: TransactionKind<L>,
-    hash: TransactionHash<L>,
-    records: &[RecordOpening],
-) -> Result<(), KeystoreError<L>> {
-    let uid = TransactionUID::<L>(hash);
-    let history = receive_history_entry(kind, records);
-    model.transactions.create(uid, history)?;
-    Ok(())
-}
-
-pub async fn add_records<
-    'a,
-    L: Ledger + DeserializeOwned + Serialize,
-    Meta: Serialize + DeserializeOwned + Send,
->(
-    state: &mut LedgerState<'a, L>,
-    model: &mut KeystoreModel<'a, L, impl KeystoreBackend<'a, L>, Meta>,
-    key_pair: &UserKeyPair,
-    records: Vec<(RecordOpening, u64, MerklePath)>,
-) -> Result<(), KeystoreError<L>> {
-    for (record, uid, proof) in records {
-        let comm = RecordCommitment::from(&record);
-        if !state.remember_merkle_leaf(uid, &MerkleLeafProof::new(comm.to_field_element(), proof)) {
-            return Err(KeystoreError::BadMerkleProof {
-                commitment: comm,
-                uid,
-            });
-        }
-
-        // Add the asset type if it is not already in the asset library.
-        model.assets_mut().create(record.asset_def.clone(), None)?;
-
-        // Mark the account receiving the record as used.
-        model
-            .sending_accounts
-            .get_mut(&key_pair.address())
-            .unwrap()
-            .set_used()
-            .save()?;
-        // Save the record.
-        model.records.create::<L>(
-            uid,
-            record.clone(),
-            key_pair.nullify(
-                record.asset_def.policy_ref().freezer_pub_key(),
-                uid,
-                &RecordCommitment::from(&record),
-            ),
-        )?;
-    }
-    Ok(())
-}
-
-pub async fn import_memo<
-    'a,
-    L: Ledger + DeserializeOwned + Serialize,
-    Meta: Serialize + DeserializeOwned + Send,
->(
-    model: &mut KeystoreModel<'a, L, impl KeystoreBackend<'a, L>, Meta>,
-    memo: ReceiverMemo,
-    comm: RecordCommitment,
-    uid: u64,
-    proof: MerklePath,
-) -> Result<(), KeystoreError<L>> {
-    for key in model
-        .sending_accounts
-        .iter_keys()
-        .collect::<Vec<UserKeyPair>>()
-    {
-        let records = try_open_memos(
-            model,
-            &key,
-            &[(memo.clone(), comm, uid, proof.clone())],
-            None,
-            false,
-        )
-        .await?;
-        if !records.is_empty() {
-            return add_records(model, &key, records).await;
-        }
-    }
-
-    Err(KeystoreError::<L>::CannotDecryptMemo {})
-}
-
-async fn clear_pending_transaction<
-    'a,
-    't,
-    L: Ledger + DeserializeOwned + Serialize,
-    Meta: Serialize + DeserializeOwned + Send,
->(
-    state: &mut LedgerState<'a, L>,
-    model: &mut KeystoreModel<'a, L, impl KeystoreBackend<'a, L>, Meta>,
-    txn: &reef::Transaction<L>,
-    res: Option<CommittedTxn<'t>>,
-) -> Option<Transaction<L>> {
-    let now = state.block_height();
-    let pending = model
-        .transactions
-        .get(&TransactionUID::<L>(txn.hash()))
-        .ok();
-    for nullifier in txn.input_nullifiers() {
-        if let Ok(record) = model.records.with_nullifier_mut::<L>(&nullifier) {
-            if pending.is_some() {
-                // If we started this transaction, all of its inputs should have been on hold,
-                // to preserve the invariant that all input nullifiers of all pending
-                // transactions are on hold.
-                assert!(record.on_hold(now));
-
-                if res.is_none() {
-                    // If the transaction was not accepted for any reason, its nullifiers have
-                    // not been spent, so remove the hold we placed on them.
-                    record.unhold().save::<L>().ok();
-                }
-            } else {
-                // This isn't even our transaction.
-                assert!(!record.on_hold(now));
-            }
-        }
-    }
-
-    // If this was a successful transaction, add all of its frozen/unfrozen outputs to our
-    // freezable database (for freeze/unfreeze transactions).
-    if let Some((_, _, uids)) = res {
-        if let Some(pending) = &pending {
-            if pending.kind().clone() == TransactionKind::<L>::freeze()
-                || pending.kind().clone() == TransactionKind::<L>::unfreeze()
-            {
-                // the first uid corresponds to the fee change output, which is not one of the
-                // `freeze_outputs`, so we skip that one
-                for ((uid, remember), ro) in uids.iter_mut().zip(pending.outputs()).skip(1) {
-                    let key_pair = model
-                        .freezing_accounts
-                        .get(ro.asset_def.policy_ref().freezer_pub_key())
-                        .unwrap()
-                        .key()
-                        .clone();
-                    model
-                        .records
-                        .create::<L>(
-                            *uid,
-                            ro.clone(),
-                            key_pair.nullify(
-                                &ro.pub_key.address(),
-                                *uid,
-                                &RecordCommitment::from(ro),
-                            ),
-                        )
-                        .ok();
-                    *remember = true;
-                }
-            }
-        }
-    }
-    pending
+    trusted_signer: &VerKey,
+    library: VerifiedAssetLibrary,
+) -> Result<Vec<AssetDefinition>, KeystoreError<L>> {
+    model.assets.verify_assets(trusted_signer, library)
 }
 
 async fn view_transaction<
@@ -1065,475 +516,123 @@ async fn view_transaction<
     Ok(())
 }
 
-async fn update_nullifier_proofs<
+async fn receive_attached_records<
     'a,
     L: Ledger + DeserializeOwned + Serialize,
     Meta: Serialize + DeserializeOwned + Send,
 >(
-    state: &mut LedgerState<'a, L>,
     model: &mut KeystoreModel<'a, L, impl KeystoreBackend<'a, L>, Meta>,
-    txn: &mut reef::Transaction<L>,
+    txn: &reef::Transaction<L>,
+    uids: &mut [(u64, bool)],
+    add_to_history: bool,
 ) -> Result<(), KeystoreError<L>> {
-    let mut proofs = Vec::new();
-    for n in txn.input_nullifiers() {
-        let (spent, proof) = model
-            .backend
-            .get_nullifier_proof(state.nullifiers, n)
-            .await?;
-        if spent {
-            return Err(KeystoreError::<L>::NullifierAlreadyPublished { nullifier: n });
+    let records = txn.output_openings().into_iter().flatten().zip(uids);
+    let mut my_records = vec![];
+    for (ro, (uid, remember)) in records {
+        if let Ok(account_editor) = model.sending_accounts.get_mut(&ro.pub_key.address()) {
+            // If this record is for us, add it to the keystore and include it in the
+            // list of received records for created a received transaction history
+            // entry.
+            *remember = true;
+            // Add the asset type if it is not already in the asset library.
+            model.assets.create(ro.asset_def.clone(), None)?;
+            // Mark the account receiving the records used.
+            let account = account_editor.set_used().save()?;
+            // Add the record.
+            model.records.create::<L>(
+                *uid,
+                ro.clone(),
+                account.key().nullify(
+                    ro.asset_def.policy_ref().freezer_pub_key(),
+                    *uid,
+                    &RecordCommitment::from(&ro),
+                ),
+            )?;
+            my_records.push(ro);
+        } else if let Ok(account_editor) = model
+            .freezing_accounts
+            .get_mut(ro.asset_def.policy_ref().freezer_pub_key())
+        {
+            // If this record is not for us, but we can freeze it, then this
+            // becomes like an view. Add the record to our collection of freezable
+            // records, but do not include it in the history entry.
+            *remember = true;
+            // Add the asset type if it is not already in the asset library.
+            model.assets.create(ro.asset_def.clone(), None)?;
+            // Mark the freezing account which is tracking the record used.
+            let account = account_editor.set_used().save()?;
+            // Add the record.
+            model.records.create::<L>(
+                *uid,
+                ro.clone(),
+                account
+                    .key()
+                    .nullify(&ro.pub_key.address(), *uid, &RecordCommitment::from(&ro)),
+            )?;
         }
-        proofs.push(proof);
     }
-    txn.set_proofs(proofs);
+
+    if add_to_history && !my_records.is_empty() {
+        add_receive_history(model, txn.kind(), txn.hash().clone(), &my_records).await?;
+    }
+
     Ok(())
 }
 
-// This function ran into the same mystifying compiler behavior as
-// `submit_elaborated_transaction`, where the default async desugaring loses track of the `Send`
-// impl for the result type. As with the other function, this can be fixed by manually
-// desugaring the type signature.
-pub fn define_asset<
-    'a,
-    'b,
-    L: Ledger + DeserializeOwned + Serialize,
-    Meta: Serialize + DeserializeOwned + Send + Send,
->(
-    state: &mut LedgerState<'a, L>,
-    model: &'b mut KeystoreModel<'a, L, impl KeystoreBackend<'a, L>, Meta>,
-    name: String,
-    description: &'b [u8],
-    policy: AssetPolicy,
-) -> impl 'b + Captures<'a> + Future<Output = Result<AssetDefinition, KeystoreError<L>>> + Send
-where
-    'a: 'b,
-{
-    async move {
-        let seed = AssetCodeSeed::generate(&mut model.rng);
-        let code = AssetCode::new_domestic(seed, description);
-        let definition = AssetDefinition::new(code, policy).context(CryptoSnafu)?;
-        let mint_info = MintInfo {
-            seed,
-            description: description.to_vec(),
-        };
-
-        model
-            .assets_mut()
-            .create(definition.clone(), Some(mint_info.clone()))?
-            .with_name(name)
-            .with_description(mint_info.fmt_description())
-            .save()?;
-
-        // If the asset is viewable/freezable, mark the appropriate viewing/freezing accounts
-        // `used`.
-        let policy = definition.policy_ref();
-        if policy.is_viewer_pub_key_set() {
-            if let Ok(account) = model.viewing_accounts.get_mut(policy.viewer_pub_key()) {
-                account.set_used().save()?;
-            }
-        }
-        if policy.is_freezer_pub_key_set() {
-            if let Ok(account) = model.freezing_accounts.get_mut(policy.freezer_pub_key()) {
-                account.set_used().save()?;
-            }
-        }
-        model.ledger_state_store.update_dynamic(state).await?;
-        Ok(definition)
-    }
-}
-
-// Add a new user key and set up a scan of the ledger to import records belonging to this key.
-//
-// `user_key` can be provided to add an arbitrary key, not necessarily derived from this
-// keystore's deterministic key stream. Otherwise, the next key in the key stream will be derived
-// and added.
-//
-// If `scan_from` is provided, a new ledger scan will be created and the corresponding event
-// stream will be returned. Note that the caller is responsible for actually starting the task
-// which processes this scan, since the Keystore (not the LedgerState) has the data structures
-// needed to manage tasks (the AsyncScope, mutexes, etc.).
-async fn add_sending_account<
-    'a,
-    L: Ledger + DeserializeOwned + Serialize,
-    Meta: Serialize + DeserializeOwned + Send,
->(
-    state: &mut LedgerState<'a, L>,
-    model: &mut KeystoreModel<'a, L, impl KeystoreBackend<'a, L>, Meta>,
-    user_key: Option<UserKeyPair>,
-    description: String,
-    scan_from: Option<EventIndex>,
-) -> Result<
-    (
-        UserKeyPair,
-        Option<impl 'a + Stream<Item = (LedgerEvent<L>, EventSource)> + Send + Unpin>,
-    ),
-    KeystoreError<L>,
-> {
-    let (user_key, index) = match user_key {
-        Some(user_key) => {
-            if model.sending_accounts.get(&user_key.address()).is_ok() {
-                // For other key types, adding a key that already exists is a no-op. However,
-                // because of the background ledger scans associated with user keys, we want to
-                // report an error, since the user may have attempted to add the same key with
-                // two different `scan_from` parameters, and we have not actually started the
-                // second scan in this case.
-                return Err(KeystoreError::<L>::UserKeyExists {
-                    pub_key: user_key.pub_key(),
-                });
-            }
-            (user_key, None)
-        }
-        None => {
-            // It is possible that we already have some of the keys that will be yielded by the
-            // deterministic key stream. For example, the user could create a second keystore with
-            // the same mnemonic, generate some keys, and then manually add those keys to this
-            // keystore. If `user_key` is not provided, this function is required to generate a
-            // new key, so keep incrementing the key stream state and generating keys until we
-            // find one that is new.
-            loop {
-                let index = model.sending_accounts.next_index();
-                let user_key = model
-                    .user_key_stream
-                    .derive_user_key_pair(&index.to_le_bytes());
-                if model.sending_accounts.get(&user_key.address()).is_err() {
-                    break (user_key, Some(index));
-                }
-            }
-        }
-    };
-
-    let (scan, events) = if let Some(scan_from) = scan_from {
-        // Get the stream of events for the background scan worker task to process.
-        let (frontier, next_event) = model.backend.get_initial_scan_state(scan_from).await?;
-        let events = model.backend.subscribe(next_event, None).await;
-
-        // Create a background scan of the ledger to import records belonging to this key.
-        let scan = BackgroundKeyScan::new(
-            user_key.clone(),
-            next_event,
-            scan_from,
-            state.now,
-            SparseMerkleTree::sparse(frontier),
-        );
-        (Some(scan), Some(events))
-    } else {
-        (None, None)
-    };
-
-    // Add a new account to our set of accounts and update our persistent data structures and
-    // remote services.
-    model
-        .sending_accounts
-        .create(user_key.clone(), index)?
-        .with_description(description)
-        .set_scan(scan)
-        .save()?;
-    model.ledger_state_store.update_dynamic(state).await?;
-    // If we successfully updated our data structures, register the key with the
-    // network. The storage transaction will revert if this fails.
-    model.backend.register_user_key(&user_key).await?;
-    Ok((user_key, events))
-}
-
-// `viewing_key` can be provided to add an arbitrary key, not necessarily derived from this
-// keystore's deterministic key stream. Otherwise, the next key in the key stream will be derived
-// and added.
-async fn add_viewing_account<
-    'a,
-    L: Ledger + DeserializeOwned + Serialize,
-    Meta: Serialize + DeserializeOwned + Send,
->(
-    state: &mut LedgerState<'a, L>,
-    model: &mut KeystoreModel<'a, L, impl KeystoreBackend<'a, L>, Meta>,
-    viewing_key: Option<ViewerKeyPair>,
-    description: String,
-) -> Result<ViewerKeyPair, KeystoreError<L>> {
-    let (viewing_key, index) = match viewing_key {
-        Some(viewing_key) => {
-            if model.viewing_accounts.get(&viewing_key.pub_key()).is_ok() {
-                return Ok(viewing_key);
-            }
-            (viewing_key, None)
-        }
-        None => {
-            let index = model.viewing_accounts.next_index();
-            let viewing_key = model
-                .viewer_key_stream
-                .derive_viewer_key_pair(&index.to_le_bytes());
-            (viewing_key, Some(index))
-        }
-    };
-
-    model
-        .viewing_accounts
-        .create(viewing_key.clone(), index)?
-        .with_description(description)
-        .save()?;
-    model.ledger_state_store.update_dynamic(state).await?;
-    Ok(viewing_key)
-}
-
-// `freezing_key` can be provided to add an arbitrary key, not necessarily derived from this
-// keystore's deterministic key stream. Otherwise, the next key in the key stream will be derived
-// and added.
-async fn add_freezing_account<
-    'a,
-    L: Ledger + DeserializeOwned + Serialize,
-    Meta: Serialize + DeserializeOwned + Send,
->(
-    state: &mut LedgerState<'a, L>,
-    model: &mut KeystoreModel<'a, L, impl KeystoreBackend<'a, L>, Meta>,
-    freezing_key: Option<FreezerKeyPair>,
-    description: String,
-) -> Result<FreezerKeyPair, KeystoreError<L>> {
-    let (freezing_key, index) = match freezing_key {
-        Some(freezing_key) => {
-            if model.freezing_accounts.get(&freezing_key.pub_key()).is_ok() {
-                return Ok(freezing_key);
-            }
-            (freezing_key, None)
-        }
-        None => {
-            let index = model.viewing_accounts.next_index();
-            let freezing_key = model
-                .freezer_key_stream
-                .derive_freezer_key_pair(&index.to_le_bytes());
-            (freezing_key, Some(index))
-        }
-    };
-
-    model
-        .freezing_accounts
-        .create(freezing_key.clone(), index)?
-        .with_description(description)
-        .save()?;
-    model.ledger_state_store.update_dynamic(state).await?;
-
-    Ok(freezing_key)
-}
-
-pub fn build_transfer<
-    'a,
-    'k,
-    L: Ledger + DeserializeOwned + Serialize,
-    Meta: Serialize + DeserializeOwned + Send,
->(
-    state: &mut LedgerState<'a, L>,
-    model: &mut KeystoreModel<'a, L, impl KeystoreBackend<'a, L>, Meta>,
-    spec: TransferSpec<'k>,
-) -> Result<(TransferNote, TransactionParams<L>), KeystoreError<L>> {
-    state.transfer(
-        &mut model.records,
-        spec,
-        &model.ledger_state_store.load_proving_keys().xfr,
-        &mut model.rng,
-    )
-}
-
-pub async fn build_mint<
-    'a,
-    L: Ledger + DeserializeOwned + Serialize,
-    Meta: Serialize + DeserializeOwned + Send,
->(
-    state: &mut LedgerState<'a, L>,
-    model: &mut KeystoreModel<'a, L, impl KeystoreBackend<'a, L>, Meta>,
-    minter: Option<&UserAddress>,
-    fee: RecordAmount,
-    asset_code: &AssetCode,
-    amount: RecordAmount,
-    receiver: UserPubKey,
-) -> Result<(MintNote, TransactionParams<L>), KeystoreError<L>> {
-    let asset = model
-        .assets
-        .get::<L>(asset_code)
-        .map_err(|_| KeystoreError::<L>::UndefinedAsset { asset: *asset_code })?;
-    let MintInfo { seed, description } =
-        asset
-            .mint_info()
-            .ok_or(KeystoreError::<L>::AssetNotMintable {
-                asset: asset.definition().clone(),
-            })?;
-    let sending_keys = match minter {
-        Some(addr) => vec![model.sending_accounts.get(addr)?.key().clone()],
-        None => model.sending_accounts.iter_keys().collect(),
-    };
-    let proving_keys = model.ledger_state_store.load_proving_keys().mint;
-    state.mint(
-        &mut model.records,
-        &sending_keys,
-        &proving_keys,
-        fee,
-        &(asset.definition().clone(), seed, description),
-        amount,
-        receiver,
-        &mut model.rng,
-    )
-}
-
-#[allow(clippy::too_many_arguments)]
-pub async fn build_freeze<
-    'a,
-    L: Ledger + DeserializeOwned + Serialize,
-    Meta: Serialize + DeserializeOwned + Send,
->(
-    state: &mut LedgerState<'a, L>,
-    model: &mut KeystoreModel<'a, L, impl KeystoreBackend<'a, L>, Meta>,
-    fee_address: Option<&UserAddress>,
-    fee: RecordAmount,
-    asset: &AssetCode,
-    amount: U256,
-    owner: UserAddress,
-    outputs_frozen: FreezeFlag,
-) -> Result<(FreezeNote, TransactionParams<L>), KeystoreError<L>> {
-    let asset = model
-        .assets
-        .get::<L>(asset)
-        .map_err(|_| KeystoreError::<L>::UndefinedAsset { asset: *asset })?
-        .definition()
-        .clone();
-    let freeze_key = match model
-        .freezing_accounts
-        .get(asset.policy_ref().freezer_pub_key())
-    {
-        Ok(account) => account.key().clone(),
-        _ => return Err(KeystoreError::<L>::AssetNotFreezable { asset }),
-    };
-    let sending_keys = match fee_address {
-        Some(addr) => vec![model.sending_accounts.get(addr)?.key().clone()],
-        None => model.sending_accounts.iter_keys().collect(),
-    };
-    let proving_keys = model.ledger_state_store.load_proving_keys().freeze;
-    state.freeze_or_unfreeze(
-        &mut model.records,
-        &sending_keys,
-        &freeze_key,
-        &proving_keys,
-        fee,
-        &asset,
-        amount,
-        owner,
-        outputs_frozen,
-        &mut model.rng,
-    )
-}
-
-pub async fn submit_transaction<
+async fn add_receive_history<
     'a,
     L: Ledger + DeserializeOwned + Serialize,
     Meta: Serialize + DeserializeOwned + Send,
 >(
     model: &mut KeystoreModel<'a, L, impl KeystoreBackend<'a, L>, Meta>,
-    note: TransactionNote,
-    info: TransactionParams<L>,
-) -> Result<TransactionUID<L>, KeystoreError<L>> {
-    let mut nullifier_pfs = Vec::new();
-    for n in note.nullifiers() {
-        let (spent, proof) = model
-            .backend
-            .get_nullifier_proof(&mut model.ledger_state_store.load_nullifiers(), n)
-            .await?;
-        if spent {
-            return Err(KeystoreError::<L>::NullifierAlreadyPublished { nullifier: n });
-        }
-        nullifier_pfs.push(proof);
-    }
-
-    let txn = reef::Transaction::<L>::cap(note, nullifier_pfs);
-    submit_elaborated_transaction(model, txn, Some(info))
-        .await
-        .await
-}
-
-// For reasons that are not clearly understood, the default async desugaring for this function
-// loses track of the fact that the result type implements Send, which causes very confusing
-// error messages farther up the call stack (apparently at the point where this function is
-// monomorphized) which do not point back to this location. This is likely due to a bug in type
-// inference, or at least a deficiency around async sugar combined with a bug in diagnostics.
-//
-// As a work-around, we do the desugaring manually so that we can explicitly specify that the
-// return type implements Send. The return type also captures a reference with lifetime 'a,
-// which is different from (but related to) the lifetime 'b of the returned Future, and
-// `impl 'a + 'b + ...` does not work, so we use the work-around described at
-// https://stackoverflow.com/questions/50547766/how-can-i-get-impl-trait-to-use-the-appropriate-lifetime-for-a-mutable-reference
-// to indicate the captured lifetime using the Captures trait.
-pub async fn submit_elaborated_transaction<
-    'a,
-    'b,
-    L: Ledger + DeserializeOwned + Serialize,
-    Meta: Serialize + DeserializeOwned + Send + Send,
->(
-    state: &mut LedgerState<'a, L>,
-    model: &'b mut KeystoreModel<'a, L, impl KeystoreBackend<'a, L>, Meta>,
-    txn: reef::Transaction<L>,
-    info: Option<TransactionParams<L>>,
-) -> impl 'b + Captures<'a> + Future<Output = Result<TransactionUID<L>, KeystoreError<L>>> + Send
-where
-    'a: 'b,
-{
-    async move {
-        let stored_txn = if let Some(mut info) = info {
-            let now = state.block_height();
-            let timeout = now + (record_root_history() as u64);
-            let uid = TransactionUID(txn.hash());
-            for nullifier in txn.input_nullifiers() {
-                // hold the record corresponding to this nullifier until the transaction is committed,
-                // rejected, or expired.
-                if let Ok(record) = model.records.with_nullifier_mut::<L>(&nullifier) {
-                    assert!(!(*record).on_hold(now));
-                    record.hold_until(timeout).save::<L>()?;
-                }
-            }
-            info.timeout = Some(timeout);
-            let stored_txn = model.transactions.create(uid, info)?;
-            model.ledger_state_store.update_dynamic(state).await?;
-            stored_txn.clone()
-        } else {
-            model.transactions.get(&TransactionUID::<L>(txn.hash()))?
-        };
-        let uid = stored_txn.uid().clone();
-        // If we succeeded in creating and persisting the pending transaction, submit it to the
-        // validators.
-        model.backend.submit(txn.clone(), stored_txn).await?;
-        Ok(uid)
-    }
-}
-
-/// Import an unverified asset.
-///
-/// Note that this function cannot be used to import verified assets. Verified assets can only be
-/// imported using [verify_assets], conditional on a signature check.
-pub fn import_asset<
-    'a,
-    L: Ledger + DeserializeOwned + Serialize,
-    Meta: Serialize + DeserializeOwned + Send,
->(
-    model: &mut KeystoreModel<'a, L, impl KeystoreBackend<'a, L>, Meta>,
-    asset: Asset,
+    kind: TransactionKind<L>,
+    hash: TransactionHash<L>,
+    records: &[RecordOpening],
 ) -> Result<(), KeystoreError<L>> {
-    assert!(!asset.verified());
-    model
-        .assets
-        .create_internal(
-            asset.definition().clone(),
-            asset.mint_info(),
-            asset.verified(),
-        )?
-        .update_internal(asset)?
-        .save()?;
+    let uid = TransactionUID::<L>(hash);
+    let history = receive_history_entry(kind, records);
+    model.transactions.create(uid, history)?;
     Ok(())
 }
 
-/// Load a verified asset library with its trusted signer.
-pub fn verify_assets<
+async fn try_open_memos<
     'a,
     L: Ledger + DeserializeOwned + Serialize,
     Meta: Serialize + DeserializeOwned + Send,
 >(
     model: &mut KeystoreModel<'a, L, impl KeystoreBackend<'a, L>, Meta>,
-    trusted_signer: &VerKey,
-    library: VerifiedAssetLibrary,
-) -> Result<Vec<AssetDefinition>, KeystoreError<L>> {
-    model.assets.verify_assets(trusted_signer, library)
+    key_pair: &UserKeyPair,
+    memos: &[(ReceiverMemo, RecordCommitment, u64, MerklePath)],
+    transaction: Option<(u64, u64, TransactionHash<L>, TransactionKind<L>)>,
+    add_to_history: bool,
+) -> Result<Vec<(RecordOpening, u64, MerklePath)>, KeystoreError<L>> {
+    let mut records = Vec::new();
+    for (memo, comm, uid, proof) in memos {
+        if let Ok(record_opening) = memo.decrypt(key_pair, comm, &[]) {
+            if !record_opening.is_dummy() {
+                // If this record is for us (i.e. its corresponding memo decrypts under
+                // our key) and not a dummy, then add it to our owned records.
+                records.push((record_opening, *uid, proof.clone()));
+            }
+        }
+    }
+
+    if add_to_history && !records.is_empty() {
+        if let Some((_block_id, _txn_id, hash, kind)) = transaction {
+            add_receive_history(
+                model,
+                kind,
+                hash,
+                &records
+                    .iter()
+                    .map(|(ro, _, _)| ro.clone())
+                    .collect::<Vec<_>>(),
+            )
+            .await?;
+        }
+    }
+
+    Ok(records)
 }
 
 // Trait used to indicate that an abstract return type captures a reference with the lifetime 'a.
@@ -1705,7 +804,7 @@ impl<
     ) -> Result<KeystoreResources<'a, L, Meta>, KeystoreError<L>> {
         let mut atomic_loader = AtomicStoreLoader::load(&loader.location(), "keystore").unwrap();
         let file_fill_size = 1024;
-        let persistence = AtomicKeystoreStorage::new(loader, &mut atomic_loader, file_fill_size)?;
+        let persistence = AtomicKeystoreStorage::new(loader, &mut atomic_loader)?;
         let adaptor = persistence.encrypting_storage_adapter::<()>();
         let ledger_state_store = LedgerStateStore::new(
             &mut atomic_loader,
@@ -1766,11 +865,11 @@ impl<
         let mut resources = Self::create_stores(loader).unwrap();
         Box::pin(async move {
             let state = if resources.persistence.exists() {
-                resources.ledger_state_store.load().await?
+                resources.ledger_state_store.load()?
             } else {
                 let state: LedgerState<'a, L> = backend.create().await?;
-                resources.persistence.create(&state).await?;
-                resources.ledger_state_store.update_dynamic(state).await?;
+                resources.persistence.create().await?;
+                resources.ledger_state_store.update_dynamic(&state)?;
                 state
             };
             resources.commit().await?;
@@ -1816,7 +915,7 @@ impl<
                 .unwrap();
         }
         Box::pin(async move {
-            resources.ledger_state_store.update_dynamic(state).await?;
+            resources.ledger_state_store.update_dynamic(&state)?;
             resources.commit().await?;
             Self::new_impl(backend, resources, state).await
         })
@@ -1978,15 +1077,17 @@ impl<
 
     /// Compute the spendable balance of the given asset type owned by all addresses.
     pub async fn balance(&self, asset: &AssetCode) -> U256 {
-        let KeystoreSharedState { model, .. } = &*self.read().await;
-        model.balance(asset, FreezeFlag::Unfrozen)
+        let KeystoreSharedState { state, model, .. } = &*self.read().await;
+        state.balance(model, asset, FreezeFlag::Unfrozen)
     }
 
     /// Compute the spendable balance of the given asset type owned by the given address.
     pub async fn balance_breakdown(&self, address: &UserAddress, asset: &AssetCode) -> U256 {
-        let KeystoreSharedState { model, .. } = &*self.read().await;
+        let KeystoreSharedState { state, model, .. } = &*self.read().await;
         match model.sending_accounts.get(address) {
-            Ok(account) => model.balance_breakdown(&account.pub_key(), asset, FreezeFlag::Unfrozen),
+            Ok(account) => {
+                state.balance_breakdown(model, &account.pub_key(), asset, FreezeFlag::Unfrozen)
+            }
             _ => U256::zero(),
         }
     }
@@ -2798,7 +1899,7 @@ async fn update_ledger<
 
 async fn update_key_scan<
     'a,
-    L: 'static + Ledger,
+    L: 'static + Ledger + DeserializeOwned + Serialize,
     Backend: KeystoreBackend<'a, L>,
     Meta: Send + DeserializeOwned + Serialize,
 >(
@@ -2847,7 +1948,7 @@ async fn update_key_scan<
         _ => false,
     };
 
-    model.ledger_state_store.update_dynamic(state).await?;
+    model.ledger_state_store.update_dynamic(state)?;
     Ok(finished)
 }
